+++ cardano-api/src/Cardano/Api.hs
@@ -593,6 +593,8 @@ module Cardano.Api (
     MempoolSizeAndCapacity(..),
     queryTxMonitoringLocal,
 
+    TxIdInMode(..),
+
     EraHistory(..),
     getProgress,
 
@@ -683,6 +685,7 @@ import           Cardano.Api.Fees
 import           Cardano.Api.GenesisParameters
 import           Cardano.Api.Hash
 import           Cardano.Api.HasTypeProxy
+import           Cardano.Api.InMode
 import           Cardano.Api.IPC
 import           Cardano.Api.IPC.Monad
 import           Cardano.Api.Key