+++ cardano-cli/ChangeLog.md
@@ -2,6 +2,10 @@
 
 ## vNext
 
+### Features
+
+- Add `query tx-mempool` ([PR 4276](https://github.com/input-output-hk/cardano-node/pull/4276))
+
 ### Bugs
 
 - Allow reading signing keys from a pipe ([PR 4342](https://github.com/input-output-hk/cardano-node/pull/4342))